requirements.txt
matplotlib
scikit-image
scikit-learn
scipy
numpy
torch
torchvision
opencv-python
opencv-contrib-python
runway-python
gdown
pillow
flask==2.0.3
waitress==2.0.0